+++ tests/cli/sensors/test_telemetry_client.py
@@ -49,6 +49,7 @@ class TestTelemetryClient(unittest.TestCase):
         self.xnames_info = [{'xname': 'x3000c0r22b0', 'Type': 'RouterBMC'},
                             {'xname': 'x3000c0s17b3', 'Type': 'NodeBMC'}]
         self.batchsize = 8
+        self.update_until_timeout = False
         self.topic = 'cray-telemetry-temperature'
         self.xname_with_data = 'x3000c0s17b3'
         self.temperature_sensors_results = [
@@ -113,9 +114,11 @@ class TestTelemetryClient(unittest.TestCase):
             ]
         }
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         self.assertEqual(self.stop_event, telemetry_client.stop_event)
         self.assertEqual(self.batchsize, telemetry_client.batchsize)
+        self.assertEqual(self.update_until_timeout, telemetry_client.update_until_timeout)
         self.assertEqual(all_topics_results[0], temperature_init_results)
         self.assertEqual(all_topics_results[1], None)
 
@@ -123,7 +126,8 @@ class TestTelemetryClient(unittest.TestCase):
         """Test stop of a TelemetryClient."""
         all_topics_results = [None]
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         self.assertFalse(telemetry_client.stopped())
         telemetry_client.stop()
         self.assertTrue(telemetry_client.stopped())
@@ -132,21 +136,24 @@ class TestTelemetryClient(unittest.TestCase):
         """Test get_topic of a TelemetryClient."""
         all_topics_results = [None]
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         self.assertEqual(self.topic, telemetry_client.get_topic())
 
     def test_endpoint_alive(self):
         """Test endpoint_alive of a TelemetryClient."""
         all_topics_results = [None]
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         self.assertTrue(telemetry_client.endpoint_alive())
 
     def test_set_sensors_for_context(self):
         """Test set_sensors_for_context of a TelemetryClient."""
         all_topics_results = [None]
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         self.assertTrue(telemetry_client.set_sensors_for_context(
                         self.xname_with_data, self.temperature_sensors_results))
         xname_results = all_topics_results[0]['Metrics'][1]
@@ -155,19 +162,84 @@ class TestTelemetryClient(unittest.TestCase):
         self.assertEqual(xname_results['Count'], 1)
         self.assertEqual(xname_results['Sensors'], self.temperature_sensors_results)
 
+    def test_set_sensors_for_context_update_existing(self):
+        """Test set_sensors_for_context where existing sensors are updated."""
+        all_topics_results = [None]
+        telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
+        self.assertTrue(telemetry_client.set_sensors_for_context(
+                        self.xname_with_data, self.temperature_sensors_results))
+        temperature_sensors_results_2 = [
+            {
+                'Timestamp': '2021-04-16T21:21:53Z',
+                'Location': 'x3000c0s17b3',
+                'ParentalContext': 'Chassis',
+                'PhysicalContext': 'Baseboard',
+                'Index': 0,
+                'DeviceSpecificContext': 'BB Inlet Temp',
+                'Value': '23'
+            },
+            {
+                'Timestamp': '2021-04-16T21:21:53Z',
+                'Location': 'x3000c0s17b3',
+                'ParentalContext': 'Chassis',
+                'PhysicalContext': 'Baseboard',
+                'Index': 0,
+                'DeviceSpecificContext': 'BB BMC Temp',
+                'Value': '27'
+            }
+        ]
+        self.assertTrue(telemetry_client.set_sensors_for_context(
+                        self.xname_with_data, temperature_sensors_results_2))
+        xname_results = all_topics_results[0]['Metrics'][1]
+        self.assertEqual(xname_results['Context'], self.xname_with_data)
+        self.assertEqual(xname_results['Type'], 'NodeBMC')
+        self.assertEqual(xname_results['Count'], 2)
+        self.assertEqual(xname_results['Sensors'], temperature_sensors_results_2)
+
+    def test_set_sensors_for_context_add_to_existing(self):
+        """Test set_sensors_for_context where new sensors are added to existing sensors."""
+        all_topics_results = [None]
+        telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
+        self.assertTrue(telemetry_client.set_sensors_for_context(
+                        self.xname_with_data, self.temperature_sensors_results))
+        temperature_sensors_results_2 = [
+            {
+                'Timestamp': '2021-04-16T21:21:53Z',
+                'Location': 'x3000c0s17b3',
+                'ParentalContext': 'Chassis',
+                'PhysicalContext': 'Baseboard',
+                'Index': 0,
+                'DeviceSpecificContext': 'Temp',
+                'Value': '30'
+            }
+        ]
+        all_results = self.temperature_sensors_results + temperature_sensors_results_2
+        self.assertTrue(telemetry_client.set_sensors_for_context(
+                        self.xname_with_data, temperature_sensors_results_2))
+        xname_results = all_topics_results[0]['Metrics'][1]
+        self.assertEqual(xname_results['Context'], self.xname_with_data)
+        self.assertEqual(xname_results['Type'], 'NodeBMC')
+        self.assertEqual(xname_results['Count'], 2)
+        self.assertEqual(xname_results['Sensors'], all_results)
+
     def test_am_i_done(self):
         """Test am_i_done of a TelemetryClient."""
         all_topics_results = [None]
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         self.assertFalse(telemetry_client.am_i_done())
 
     def test_unpack_data(self):
         """Test unpack_data of a TelemetryClient."""
         all_topics_results = [None]
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
-
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         num_metrics = telemetry_client.unpack_data(json.dumps(self.event_data))
         self.assertEqual(num_metrics, 1)
         self.assertFalse(telemetry_client.am_i_done())
@@ -177,11 +249,45 @@ class TestTelemetryClient(unittest.TestCase):
         self.assertEqual(xname_results['Count'], 1)
         self.assertEqual(xname_results['Sensors'], self.temperature_sensors_results)
 
+    def test_unpack_complete_data(self):
+        """Test unpack_data with all sensors for 1 xname."""
+        all_topics_results = [None]
+        xnames_info = [{'xname': 'x3000c0s17b3', 'Type': 'NodeBMC'}]
+        telemetry_client = TelemetryClient(self.stop_event, xnames_info, self.batchsize,
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
+        num_metrics = telemetry_client.unpack_data(json.dumps(self.event_data))
+        self.assertEqual(num_metrics, 1)
+        self.assertTrue(telemetry_client.am_i_done())
+        xname_results = all_topics_results[0]['Metrics'][0]
+        self.assertEqual(xname_results['Context'], self.xname_with_data)
+        self.assertEqual(xname_results['Type'], 'NodeBMC')
+        self.assertEqual(xname_results['Count'], 1)
+        self.assertEqual(xname_results['Sensors'], self.temperature_sensors_results)
+
+    def test_unpack_complete_data_update_until_timeout(self):
+        """Test unpack_data with all sensors for 1 xname with update_until_timeout=True."""
+        all_topics_results = [None]
+        xnames_info = [{'xname': 'x3000c0s17b3', 'Type': 'NodeBMC'}]
+        update_until_timeout = True
+        telemetry_client = TelemetryClient(self.stop_event, xnames_info, self.batchsize,
+                                           update_until_timeout, self.topic,
+                                           all_topics_results, 0)
+        num_metrics = telemetry_client.unpack_data(json.dumps(self.event_data))
+        self.assertEqual(num_metrics, 1)
+        self.assertFalse(telemetry_client.am_i_done())
+        xname_results = all_topics_results[0]['Metrics'][0]
+        self.assertEqual(xname_results['Context'], self.xname_with_data)
+        self.assertEqual(xname_results['Type'], 'NodeBMC')
+        self.assertEqual(xname_results['Count'], 1)
+        self.assertEqual(xname_results['Sensors'], self.temperature_sensors_results)
+
     def test_check_if_run_done_when_stopped(self):
         """Test check_if_run_done of a TelemetryClient that is stopped."""
         all_topics_results = [None]
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         self.assertFalse(telemetry_client.check_if_run_done(self.topic))
         telemetry_client.stop()
         self.assertTrue(telemetry_client.check_if_run_done(self.topic))
@@ -190,7 +296,8 @@ class TestTelemetryClient(unittest.TestCase):
         """Test ping_and_sleep of a TelemetryClient."""
         all_topics_results = [None]
         telemetry_client = TelemetryClient(self.stop_event, self.xnames_info, self.batchsize,
-                                           self.topic, all_topics_results, 0)
+                                           self.update_until_timeout, self.topic,
+                                           all_topics_results, 0)
         self.assertTrue(telemetry_client.ping_and_sleep(self.topic))